requirements/default.txt
boto3==1.7.67
dateparser==0.6.0
environs==2.1.1
Flask-Caching==1.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.1
Flask==1.0.2
gevent==1.3.5
gunicorn==19.9.0
humanize==0.5.1
jmespath==0.9.3
jsonschema==2.6.0
kafka-python==1.4.3
lxml==4.2.4
MySQL-python==1.2.5
protobuf==3.6.0
python-dateutil==2.6.0
python-jose==3.0.0
pytz==2017.2
redis==2.10.6
requests==2.19.1
sendgrid==3.6.3
six==1.11.0
slackclient==1.2.1
validate-email==1.3
kafka-python==1.4.3
mysql-replication==0.18
requests-aws4auth==0.9
elasticsearch==6.3.1
